Refactor some preferences functionality into preferences module
