+++ actions.py
@@ -1,5 +1,4 @@
 import datetime
-from sqlite3 import IntegrityError
 from typing import List
 
 import pandas as pd
@@ -14,7 +13,6 @@ import auth
 from flask_mail import Mail
 
 from auth import generate_confirmation_token
-from preferences import DefaultPreferences
 
 mail = Mail()
 
@@ -46,24 +44,6 @@ def add_tomorrows_forecast_to_db(location: models.Location):
 
 def update_preferences(form):
     print(form.data)
-
-
-def load_user_preferences(form, user):
-    """Load user preferences"""
-    return form
-
-
-def create_default_preference_row(preferences=DefaultPreferences):
-    return models.Preferences(day_start=preferences.day_start,
-                              day_end=preferences.day_end,
-                              temperature=preferences.temperature,
-                              activities=preferences.activities)
-
-
-def assign_default_preferences(user):
-    """Create a new row in the preferences table for `user` and populate it with default values"""
-
-    return models.Preferences(user=user)
 
 
 def register_new_user(email: str, place: str):
